# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Productos",
        "description": "API para la gestión de productos y precios en diferentes divisas. TEST CIPH3R",
        "contact": {
            "email": "mperez.392@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obtener la lista de productos",
                "operationId": "080f057905dc32d2f11065d606dc4cd6",
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
        "/api/products": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Crear un nuevo producto",
                "description": "Crea un producto con su precio en la divisa base y otros precios en diferentes divisas.",
                "operationId": "createProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "currency_id",
                                    "prices"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Producto A"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción del Producto A"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "currency_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tax_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 20
                                    },
                                    "manufacturing_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "prices": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "currency_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 120
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud inválida"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obtener un producto por ID",
                "description": "Este endpoint permite obtener un producto específico usando su ID.",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto que se desea obtener",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto encontrado y retornado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Actualizar un producto y sus precios",
                "description": "Este endpoint permite actualizar los detalles de un producto, incluyendo los precios en diferentes divisas. Los precios existentes se sincronizan, permitiendo la eliminación de precios de divisas no incluidas.",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "currency_id",
                                    "prices"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Producto A"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción del Producto A"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "currency_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tax_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 20
                                    },
                                    "manufacturing_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "prices": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "currency_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 120
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "400": {
                        "description": "Solicitud inválida"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Eliminar un producto",
                "description": "Este endpoint permite eliminar un producto y sus precios asociados.",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto eliminado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "400": {
                        "description": "Solicitud inválida"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Modelo de Producto",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Producto 1"
                    },
                    "description": {
                        "type": "string",
                        "example": "Descripción del producto"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    },
                    "currency_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tax_cost": {
                        "type": "number",
                        "format": "float",
                        "example": 15
                    },
                    "manufacturing_cost": {
                        "type": "number",
                        "format": "float",
                        "example": 30
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}